attmpt img context hangs
- ensure png is jpeg
- prev error: Error: API Error: messages.0.content.1.image.source.base64.data: Image does not match the provided media type image/jpeg
- now hangs: Fetching 2 files: 100%|██████████| 2/2 [00:51<00:00, 25.82s/it]
Loading checkpoint shards: 100%|██████████| 2/2 [00:02<00:00,  1.23s/it]
Using a slow image processor as `use_fast` is unset and a slow processor was saved with this model. `use_fast=True` will be the default behavior in v4.52, even if the model was saved with a slow processor. This will result in minor differences in outputs. You'll still be able to use a slow processor with `use_fast=False`.